Some types in import_compilations are not longer templated.
Some of the lower level types defined by import_compilations.hpp don't
need to be templated on a transaction: a database reference is sufficient.
