+++ lib/exchange/CMakeLists.txt
@@ -79,6 +79,7 @@ set (pstore_exchange_sources
     export_fragment.cpp
     export_names.cpp
     export_ostream.cpp
+    import_compilations.cpp
     import_error.cpp
     import_fixups.cpp
     import_fragment.cpp
